requirements.txt
fastapi[standard]
PyPDF2
python-docx
python-multipart
openai
python-dotenv